requirements.txt
langchain
langgraph
langsmith
python-dotenv
langchain-groq
langchain_tavily
langchain-mcp-adapters
mcp
langgraph
langgraph-cli[inmem]
opik